revart unchangeable props to changeable props
My initial understanding was if set method is excluded,
JS will throw error when someone tries to modify the property value.
But this is only TS error, not JS error. JS being JS,
will not throw any error, only it will not update any value.
This can lead to unexpected behavior on codes using the sdk.
So removed the get methods where set method is not defined explicitly.